Filling out unit tests
Here I have added unit tests to the materials area.
These tests are not creative or original, and where possible I have tried
to verify the numbers using various sources online.
One reason to add unit tests to this PR is to increase the code coverage
of the PR as a whole, which dropped due to removing lines of code.