Another Pylint PR (#294)
* Removing useless pylint code duplication detection

* Removing circular imports

I found some circular imports all over the code base most stemming from this line:

import armi

So, lots of our base code still has this, as it will take significant redesign of
the global package variables and setting stuff. However, some of the instances
were easy to fix. Particularly the unit tests.

I should note that Python does a good job of avoiding circular imports, so most
of these weren't a REAL problem. However, it is probably bad practice to use
import armi in our code base, basically anywhere.

* Pylint cleanup - C0113 correct use of not in

* Pylint cleanup - R1714 consider-using-in

* Pylint cleanup - W0612 unused-variable

* py2to3 - removing future print import

Several files still had this line at the top:
from __future__ import print_function
Generally, this code is only useful in Python v2.

There are only a tiny amount of edge cases where
this line is useful in Python v3, mostly having
to do with trying to connect to Python v2 modules.
But it doesn't appear to me this is the case.

* pylint cleanup - W0612 unused-variable

* Pylint cleanup - R1711 useless-return

* Addressing PR comments

Mostly, this commit addresses the renaming/removing of unused variables.
But this also addresses some conflicts between the branches.

* Filling out unit tests

Here I have added unit tests to the materials area.
These tests are not creative or original, and where possible I have tried
to verify the numbers using various sources online.
One reason to add unit tests to this PR is to increase the code coverage
of the PR as a whole, which dropped due to removing lines of code.