Make it possible to add interfaces with the same function (#307)
Originally addInterface would raise an exception when the Interface
stack already contained an Interface with the same function as the one
passed into it. This was limiting when you had different plug-ins but
wanted the right Interface to take precedence.
So, now it adds an Interface even if it has the same function as one
previously if the new Interface is a subclass of the current Interface
with that function. This keeps the Interface that is more derived and
removes the old Interface. This still throws an exception if the new
Interface has the same function but shares no inheritance with the old
Interface. If the old Interface is instead a subclass of the new
Interface, neither interface is removed or added.