Add option to grid GUI to save map to an image (#313)
It is common to want to get at an image of the grid layout for documentation
purposes. Until now one could take a screenshot or similar. However, often times
the grid is too large to fit on the screen, requiring the user to take multiple
screen shots at different scroll positions and stitch the images together. Yuck.
This adds a button for saving the whole grid layout as one image.