Add expand-bp entry point
This entry point will process the !include directives in an ARMI YAML
file, emitting the results to standard output. This also adds a class
attribute the the EntryPoint class to control whether a splash text
should be writen to stdout, allowing the output to be safely captured in
a new YAML file.