Improve plotting behavior.
Adjust plotFaceMap where _setPlotValText now takes in the collection
currently used for plotting so that we have a reference to what colors
were used for each Hex. In passing in the collection, we get access to
its color map and norm so we can go back and find what patch color went
with what value originally. With this information we can figure out if
it is a dark or a light color and assign text to be white if a dark
patch and black if a light patch.

In addition to this, also, changed colorbar to be fig.colorbar and not
ax.colorbar since ax.colorbar was not working and was getting
AttributeError: 'AxesSubplot' object has no attribute 'colorbar' so I
looked into it and I think we want to add the colorbar to the fig and
not the ax.

Also, for plotFaceMapLegend, just passed in the collection instead of
the norm and cmap separately and also added a size argument that
defaults to 9 (the original legend size for faceMaps) since this could
make it more generalized for use to create different legends later on that may
need to be a different size.