Update copyInterfaceInputs() to use glob.glob() (#324)
than pathlib.glob. For Windows environments with Python 3.7.2 the
pathlib.glob does not preserve the case sensitivity of file names when
it finds a match. This is problematic when interfaces are copying python
scripts (like fuel management, or custom control rod logic) since the
imports of these scripts is case sensitive.