Change how assembly pitch is found
Originally, would take the average of assembly pitches and return that
as the assembly pitch. Ran into a problem when using blueprints from a
Thermal Reactor and it did not like how originally it called "pitches =
numpy.array([b.getPitch() for b in self.getBlocks(Flags.FUEL)])" but in that specific case, no flags were fuel so pitches would be empty. But now, since we have spatialGrids, just calling
self.spatialGrid.pitch may do what we want as well and get over this no
blocks with a certain flag issue.