Create Grid and Locators for a Block (#323)
In blocks, creates autoCreateSpatialGrids() that when called upon for
a block with no spatialGrid currently, will create one for it if it
meets certain conditions. These conditions are that among the blocks
components, the multiplicities are either equal to 1 or N. (So there can
be no greater than two multiplicities among the components and one of
the multiplicities must be 1). Things with mult equal to 1 will have a
spatialLocation located at the center of the grid. Those with
multiplicity N will have a MultIndexLocation and appear at every pin
position location in the grid (they will appear N times). The
functionality of this is for the added feature of Block Diagrams.

In blockBlueprints, makes it so that at construction, a block is given a
spatialGrid and its components spatialLocators if certain conditions are
met. These conditions are that it is a HexBlock (meaning the function
autoCreateSpatialGrids within blocks.py is not implemented for non-Hex
Blocks, and an exception is caught and the grid remains none) and it has
basic multiplicities as defined within autoCreateSpatialGrids (mult = 1
or N, only two possible multiplicities). This sets up blocks so that
they may later be used to create basic block diagrams in cases where the spatialGrid was
not originally provided but the component placement could be inferred.