Make Diagrams use Hollow Hexes and Squares.
Made it so that the patches made for plotting Block Diagrams were
hollowed and not overlapping (as they were before). The Derived Shape
element is now dependent on the size of the larges element and placed
down first so that every other element is placed on top of it.

This is accomplished by representing Hexagons as 6 polygons (or squares
as 4 polygons) as oppossed to a single shape. This takes into account
the inner and outer pitch or diameter of the shape and allows it to appear
cut out in the center.

In addition, changed the type of error in both _makeBlockPinPatches and
_makeAssemPatches to TypeErrors when ThetaRz grids are passed in.

Also, in autoCreateSpatialGrids, had to change the indexLocation
assignment since for some blocks, it was not pairing the grid with the
components indexLocation and was saying that the indexLocation had not
grid. Now it should not do that.

Also added some test cases to test_plotting.py as well as changed
_createFaceMap legend to be _createLegend to more generally handle both
FaceMap legend and Block Diagram legend creation. It now takes in a
shape keyword in order to allow Block Diagrams to not use Hexagons, now
they currently use rectangles.