Add ability to plot Block Diagrams (#329)
Create a function plotBlockDiagram in plotting.py in order to plot block
diagrams when a spatial grid is on the block. In addition, the
_makeBlockPinPatches and _makeComponentPatch work to create component
patches as long as the block's grid type is supported.

For each components spatialLocator, it creates a patch for the
components shape and size at the location/locations described by the
locator and colors according to material name the component is made out
of. Also creates a legend saying what color is what material. At the
end, plotBlockDiagram saves the figure as an SVG.