Make material color the same across Block Diagrams
Originally, the same material would result in a different color in a
diagram for a different block within the same report (For example, Sodium would be
Red in one image, but Yellow in another). This change adds the ability
to pass in a list of all materials that would be within all of these
block diagrams, so that if you were to plot multiple block diagrams that
had the material Sodium, Sodium would always be the same color in each
diagram.

Also, got rid of bbox_to_anchor argument for the pltkwargs of
saveFig since it had the warning "MatplotlibDeprecationWarning:
savefig() got unexpected keyword argument "bbox_to_anchor" which is no
longer supported as of 3.3 and will become an error two minor releases
later", and removing it still allowed for the legend to be centered
at the right hand side of the Diagrams.