Allowing submodules to import and use the Python logger.
This was accomplished by creating a subclass of Logger, RunLogger,
that can handle some of our logging excentricities. It also involved
creating a logging Filter to allow users to stop duplicate warnings
from swamping the logs.

There was a minor change to our pytest.ini here to allow for unit
testing with PyText fixtures in our pytest-xdist parallel
environment. This works fine, but oddly means that even when
running unit tests NOT in parallel, we invoke `-n 1` as an
argument.