Updating grammar and documentation errors in the SFR tutorial (#345)
* Add metadata for SFR example in tutorial docs

* In the tutorial section of the docs, there is an example for building a
full SFR input. The step-by-step process outlined for building this
input does not include the necessary metadata parameter in the settings
file (it is noted that the metadata _is_ included in the provided
settings file which may be downloaded). Since the tutorial says that
"You have now created all the ARMI inputs, from scratch, needed to
perform a simplified analysis...", it is important that the input
that the user creates from following the instructions actually works.

* This has been fixed by adding an instruction in the tutorial for the
user to write the needed metadata line. In addition, the yaml anchor
needed to properly build the docs has been added to the provided
settings file.