Make changes in structure to reports
Changed reports from a dictionary of dictionaries to a dictionary of
Sections. These Sections are themselves ReportNodes and contain a render
function allowing the ability to transfer the contents into htmltree
elements. Also, can now add htmlelements directly into the tree.
In addition to this, this changed the layout of how other
things are accessed/added to the reports so the other files were changed
according to this new structure.