Fix tutorial documentation associated with fluxSolver and thermalSolver
This commit specifically fixes the heat balance equation represented as part of
the `thermalSolver` script in the 'making your own app' tutorial. This is done
in tandem with another fix in the [armi-example-app repo](https://github.com/terrapower/armi-example-app/pull/1)
which updates the psuedocode in the comments of `thermalSolver` to be consistent
with the changes of this commit.