Make Tables, Images, TimeSeries without title.
Originally, these items would have there own large title as if they
themselves were Section objects. Now we only have that if the item is
within a section itself. Tables still do have titles, but this happens
in its own rendering. This is because Tables, Images, Timeseries are not
Sections, but elements that would be within sections.

In addition, also made __contains__ a method of Section, and cleaned up
the example within TimeSeries.