Add new report structure and it's documentation
Add a new way of compiling report content. In general, ReportContent is
the way report content is stored. It is made up of ReportNodes(Sections, Tables,
Images, TimeSeries(the storage representation for eventual graph output) and HtmlElements. It is converted to an html page
output when writeReports() is called and renders each of it's items into
html elements. In addition, getReportContents() is a plugin hook that
collects contents from different plugins over the course of collecting
data for the report. This allows for different plugins to either add to
an existing Section/Table/etc or create new Section/Tables/etc

Currently this collection happens in the entry point for
reports, that allows a user to either provide a settings yaml file or an
h5 file to create report content.

In addition, created developer documentation for these items within
reports.rst. Also, added added a section for accessing the report entry
point within the User Documentation section.