Moving to faster YAML reads by default
The Yamlize library defaults to using a read-and-write YAML library
to read: ruamel.yaml.RoundTripLoader, but this is ~12x slower at
reading large YAML files than CLoader, which is best for situations
where you are only reading the YAML.

So, we are using the Yamlize API to choose the faster YAML Loader
library, in situations where we are only reading YAMLs. This is an
easy win for performance, as long as we don't accidentally miss
a situation that is not strictly read-only.