Refactor runLog to use logging standard library (#336)
The goal of refactoring the runLog is to use more standard Python
library tools, and to make it less MPI-aware. For backwards
compatibility reasons, this version still supports the previous
runLog API:

1. building custom logging strings
2. adding custom strings to stdout in the parent process
3. not touching stderr in the parent process
4. using logging to pass child stdout to log files
5. using logging to pass child stderr to log files
6. concatenating all logs in a custom way after the run
7. tracking all warnings, and counting their occurences
8. prevent duplicate warnings
9. providing a way to print all warnings after a run
10. create log directories
11. allowing users to change the global stdout log level

This was accomplished by creating a subclass of Logger, RunLogger,
that can handle some of our logging excentricities. It also involved
creating a logging Filter to allow users to stop duplicate warnings
from swamping the logs.

There was a minor change to our pytest.ini here to allow for unit
testing with PyText fixtures in our pytest-xdist parallel
environment. This works fine, but oddly means that even when
running unit tests NOT in parallel, we invoke `-n 1` as an
argument.