Update to running directory in the generic ExecutionOptions class. (#363)
* Updating the default run directory within the generic ExecutionOptions
to be based on the first eight characters of the hash of combining
the case title and the label. This helps to create a unique directory
name for running the case but also helps to avoid OS related file path
length issues (particularly in Windows).

* Adding unit tests for the ExecutionOptions class to test the
run directory path. As a result, changing from the basic hash()
function to hashlib since the hash() function is non-deterministic
with Python 3 without explicitly setting the PYTHONHASHSEED
environment variable.