Add Report Content Structure (#361)
This adds classes and functions that support the creation and manipulation of
report contents. Most of these are documented in doc/developer/reports.rst. This
also implements some basic report contents for reactor design summary, case
summary and neutronics.

This also introduces the getReportContents() plugin API, which allows plugins to
advertise report contents to ARMI, and a `report` entry point for generating
reports from either a settings file, or a database file.