* Add start of runLog to the case.py so that the verbosity can be set when a run begins versus waiting until an operator is created.
* Update to the `specifyInputs` labels in fuel handlers so that the
stdout has the correct labels for shuffle logic and explicit
shuffles files.

* Add more runLog messages around checking if the lattice physics
interface will create XS for the current cycle. This helps a user
to debug what is going on.

* Add check for `SameFileError` exception during the `copyOrWarn`
path tools function. This skips providing a user warning if a
file is being copied to the same path and just lets it happen
silently.