Deploying Docs to GitHub Pages using GitHub Actions
Adding a GitHub Actions YAML file to start a Workflow that builds
the ARMI docs and deploys them to the `gh-pages` branch in the repo.
This will auto-deploy the latest docs, every time a change is made
to the `master` branch.

To aid in this goal, there is a new `tox` command to build our docs:
`tox -e doc`.