Add a DeprecationWarning that settings will be immutable soon. (#374)
* Adding a DeprecationWarning that settings will be immutable soon.

Here is the little experiment in response to #371

There are a lot of decisions that went into this experiment that people
might question. For instance, I fixed a TON of unit tests so they wouldn't
throw the deprecation warning. Also, the `lock` I put in the `Settings`
class is very simple. Just decisions.

This experiment also shows us that `reactors.py` and `database3.py` will
need to be slightly re-designed to not change settings during a run.

This PR would serve as a good warning to people working on plugins that
a big change is on it's way.