Changing from github.actor to github.repository_owner
The problem this commit is trying to solve is one of forked repos.
We want the GitHub Actions to pass on the terrapower account for
armi and terrapower.github.io.
But if a developer forks both of these repos, they should be able
to have GitHub auto-run all the Actions for them, without incident.
Which means GitHub has to use "terrapower" as the user name some
times, but some times it should use "john-science", if that user
has forked both repos.