Add support for explicit Case dependencies (#408)
This is in response to #400. We probably want to ultimately get to a place where Case dependencies are managed from the Suite side, but this should compose well with that (case.addDependency() could still function by interacting under the hood with it's parent Suite), and solves the immediate issue of supporting simple, explicit Case dependencies.