Ditching GitHub's internal cache of our code coverage badges
This is a known issue with GitHub caching badges of all kinds.
Adding any query garbage to the end of our badge URL will force
GitHub to grab a new copy of our badge. Which means any time our
code coverage changes substantially we should rev the kill_count
in this URL.
It appears GitHub is defeating the whole purpose of these badges,
but there isn't much we can do about that for now.