Remove BlockNumberModifier.
This class was not tested and didn't work anymore. Attempts at a unit
test ran into a NotImplementedError from the block.duplicate() call.
Fixing would require defining the requirements of the .convert() method
more precisely, which aren't really clear. We can bring it back later
if we know exactly what we want this to do and how we want to do it.