Warning when unused material modifications are present
When the user puts a material modification in a YAML file (or in code), they
expect them to be used. In the past, a ficticous (or unused) mat mod would
be passed silently, and no one would know. It seems greatly preferrable that
users be informed that their run probably wasn't exactly what they expected.
This commit only modifies fuel-like materials.