Add new top-level components input for blueprints. (#505)
This adds a new `components` section to the top-level
blueprints input file. Components defined there are loaded
into the new `blueprints.componentDesigns` data structure,
where they can be found and used to construct components.

This is a new and optional way to define components, separate
from the traditional way of defining them within parent blocks.

Note that with this commit, there is no way to automatically
load components into blocks, etc. That would have to be done
programmatically, for now.

This is related to a desire to have more flexible ways to define
models, including e.g. TRISO fuel, MSR models, etc.