Fixed bug where local commit hash fails in a fork with no tags.
The staticmethod `getLocalCommitHash()` tries to build a uniquely
identifying string from the local commit. It's a good idea. But if
you are running this method in a forked repository, and that repo
has no commit tags, this method fails.
However, the method obviously intends to return "unknown" in this
case, as that is the standard `git describe` error code. I have
fixed the method to do so.
This commit closes issue #509.