Add the ability to specify material modifications on a component level
This adds the ability to specify material modifications on a component level
within a block. Using the `componentwise` attribute in the `material modifictions`
section of the blueprints, a given material modification may be specified to apply
to only a selection of components in the block.

Backwards compatibility is maintained in terms of the blueprints input file.
But an additional attribute `blockwise` has been added for the user to highlight
that a given material modification applies to all applicable components in the block.
This `blockwise` keyword may be used optionally and material modifications without
either the `componentwise` or `blockwise` keyword will be treated blockwise as
before.

The user docs have been updated to reflect this change.