Remove the `_compileAssemblyMaterialInput` method and retool associated funtionalities
The `_compileAssemblyMaterialInput` method was duplicating data unnecessarily, so
it was gotten rid of. To get rid of it, some minor changes needed to be made, like
changing the material modification yamlize defaults and doing some dict manipulation
on the fly.

In addition, the `blockwise` yamlize attribute was removed. Now if the user wants
to specify a `blockwise` material modification, they do so using the old syntax.
This will be reflected in the docs in a future commit.