Auto-correct ring schedules out of bounds.
This now adjusts ring schedules that are beyond the edge of the core
and crashes if the jumpRing is outside the core. New tests were added to
demonstrate the expected behavior more rigorously. This may impact some
external shuffling inputs, but if it does those may not be doing what
the user expected.