Add basic support of component groups.
This allows the user to input component groups intended to mix different
free components in different fractions. It is a follow-up to #505 as
part of the larger #504 task related to improving input flexibility.

This implements the addition of groups of components to blocks, which
can be used to create more complex and flexible models.

The code changes solved a few residual issues where iterating over the
children of a block were assuming all children would be Components. Now
they can be Composites with multiple Components or Components.

Some upgrades to DerivedShapes were necessary so that they could work
with volumetric components in addition to pure 2D ones.

These are additional steps toward #504, but true practical usage of
these groups is not quite done yet. More followups are needed.

Added a new test covering negative volume in derived shapes