Fixing the Setting copy to also copy non-default values. (#534)
It appears that a recent PR removed a single line of code from setting.py,
this has lead to a invisibly broken system that works for nearly all unit
tests. I have added a unit test to explicitly check for this important
feature in the future.
Also, bumped the ARMI version to 0.2.1.