Updates to fix latex build of docs.
The LaTeX builder did not work with SVGs so I converted them to pngs. I
also fixed some incorrect LaTeX math and dealt with a few stray
backslashes that were failing the build.

With these changes, the Sphinx LaTeX builder works properly and xelatex
builds a full pdf appropriately.

It's not a perfect PDF in that the introductory sections are added as
top level chapters to the end of the document, but other than that it is
pretty good.