Componentwise mat mods (#517)
* Fix applyInputParams bug when values are zero

* Only check against user-provided power when the current node power is available

No new capability is added. The check of the power on the uniform mesh against the
user-provided power for the cycle is only changed so that it no longer occurs on
the BOC time node of each cycle, when the power data on the blocks is stale.

* Add the ability to specify material modifications on a component level

This adds the ability to specify material modifications on a component level
within a block. Using the `componentwise` attribute in the `material modifictions`
section of the blueprints, a given material modification may be specified to apply
to only a selection of components in the block.

Backwards compatibility is maintained in terms of the blueprints input file.
But an additional attribute `blockwise` has been added for the user to highlight
that a given material modification applies to all applicable components in the block.
This `blockwise` keyword may be used optionally and material modifications without
either the `componentwise` or `blockwise` keyword will be treated blockwise as
before.

* Remove the `_compileAssemblyMaterialInput` method and retool associated funtionalities

The `_compileAssemblyMaterialInput` method was duplicating data unnecessarily, so
it was gotten rid of. To get rid of it, some minor changes needed to be made, like
changing the material modification yamlize defaults and doing some dict manipulation
on the fly.

In addition, the `blockwise` yamlize attribute was removed. Now if the user wants
to specify a `blockwise` material modification, they do so using the old syntax.
This will be reflected in the docs in a future commit.