Fix broken rendering of badges on GH.
In #539 we made some fixes to try to get LaTeX and HTML documentation
builds to both work. In doing so, the fix we made broke the badge
rendering on the main landing page of the repo on Github. Thus, this is
a different fix for that issue.

This uses a sphinx plugin to automatically convert the SVGs to pngs
before rendering and removes the .. only:: directive from the README.