Provide ArmiPlugin.defineSystemGrid builders
Goal is to let a plugin define additional sections that
can reside in the ``systems`` section, e.g.,
```yaml
systems:
    core:
        grid name: core
        origin:
            x: 0
            y: 0
            z: 0
    sfp:
        grid name: spent fuel pool
        type: sfp
        ...
```

The end goal is to have a new system for control drums, pointing to a designated
grid, and allowing assemblies to be added to that grid rather than the core grid.

The plugins basically define a map from the system grid type (defaults to "core"
in the blueprint) and the builder function. The builder function will most likely
be the class of the thing to be built (e.g., ``"core"`` returns a ``Core`` **type**
and the blueprint construct method creates the instance given the name of the grid.
In the core example it's not super helpful. But for the spent fuel pool, the
``type`` is ``sfp`` but the name is defined by the key one level up and does not
have to match any specific criteria.

The `ReactorPlugin` defines the existing ``"core"`` and ``"sfp"`` builders, but is
tagged to go last (or near last) in the YAML construction sequence. This allows other
plugins to define their own core builders and let those have preference over the
default.

Signed-off-by: Andrew Johnson <a.johnson@usnc-tech.com>