Block.r uses getAncestor if Block.core is None
Found for some objects that exist as siblings to the Core in the Reactor,
various HDF serialization methods would error out because Block.core -> None.
One example:
_addHomogenizedNumberDensityParams -> collectBlockNumberDensities ->  getSymmetryFactor
calls self.r.core.spatialGrid.overlapsWhichSymmetryLine

The docstring for Block.r indicates that most calls to self.r are to obtain the core
and may be better off just calling Block.core. But, here exists a case where a Block
may not have a Core as the parent (e.g., ex-core structures that physically reside
in the reactor) but still needs to access, for some reason, the spatial grid.

A deeper issue is that this call stack is trying to find the spatial grid for a block
assuming the block is on the Core.spatialGrid. Not sure how to square that. Maybe
the call for symmetry line tries to find a parent with a spatialGrid?

Signed-off-by: Andrew Johnson <a.johnson@usnc-tech.com>