Fix to a bug where an ARMI application could not call (#593)
`armi.configure(app, permissive=True)`. This error is
due to fact that the type check for the application that
is passed in would never be equal to the type of the base
ARMI app class. This now checks if the application passed
in is instead an instance of the ARMI app.