Fix reading records 2D and 4D, make notes for record 3D
Record 2D for VARIANT NHFLUX contains different data than Nodal NHFLUX.
The VARIANT version includes extra data at the end of the record. A fix
is added to address this mechanically, but the data is discarded.

Record 3D for VARIANT NHFLUX has a different meaning than Nodal NHFLUX.
The VARIANT version represents coefficients of the polynomial basis for
both the even and odd parity fluxes. The Nodal, version, on the other
hand, has no concept of even and odd parity. For P1 transport with P0
scattering, the existing code still works for VARIANT, but it will break
for everything else.

Record 4D for VARIANT flux, due to the changes on Record 2D, also
required a change. The existing code still ignores the remaining VARIANT
moments.