Fixing assembly number issue in DB Load (#615)
It appears there has been a bug in our DB load for several years now. Essentially, when reading from certain (though not all) H5 databases, the assembly number can start off in a bad state. This is not immediately obvious, but if you try to do something serious with the resultant Reactor/Core objects, you immediately find the problem. In this case, trying to use growToFullCore() immediately after the load() raises the problem.

NOTE: Relatedly, it turns out Block.makeUnique() was unused. So I removed it. Thanks @ntouran !