Update the handling of the various cycles case settings
This is quite a big change that updates the parsing, defaults, validation, and
setting of the cycle-related case settings. This includes `cycleLength(s)`,
`availabilityFactor(s)`, `burnSteps`, and `powerFractions`, as well as the new
setting `cycles` for the detailed cycles format.

All defaults for these settings have been removed. The user must now enter either
a complete set of the simple cycles inputs or the `cycles` detailed input.

THIS WILL BREAK ANY INPUTS THAT HAVE BEEN RELYING ON THE CYCLES DEFAULTS!

If both are entered, the user is prompted to use the detailed inputs. If only a partial
set of the simple settings are entered, the user is prompted to set the remaining
simple settings to what was previously their default values. If the detailed
`cycles` format is used, the various simple cycles inputs are all reverted to
`None`s. This _forces_ code that wants this information to access it through
the operator so that accidental mixing of the simple/detailed inputs can't happen.

In addition, the resolution of these settings on the operator has been changed
to add flexibility to how the values can be accessed. To get the values, the user
can still pull them from the operator. But they can also pull it directly out of the
case settings via `operator.getBurnSteps` or whatever, which will be accessible
even if the operator is not initialized. Further, the different cycles settings
are resolved one at a time at the time of first access so that, if for whatever
reason (probably during a unit test) all values are not input, they existing
ones can still be accessed.

Finally, the places in which these changes broke the framework have been fixed.

Updated unit tests to come.